+++ backend/src/index.ts
@@ -13,6 +13,7 @@ import uploadRoutes from "./routes/upload";
 import downloadRoutes from "./routes/download";
 import notificationRoutes from "./routes/notification";
 import notificationSettingRoutes from "./routes/notification-settings";
+import supportTicketRoutes from "./routes/support-ticket";
 import configRoutes from "./system/config";
 
 import logger from "./utils/logger";
@@ -32,11 +33,13 @@ import {
 import autoMailRoutess from "./system/automail";
 import backupRoutes, { listCollection } from "./system/backup";
 import versionRoutes from "./system/version";
+import path from "path";
 
 const DEFAULT_PORT = 3000;
 const port = Number(process.env.PORT) || DEFAULT_PORT;
 
 const HOST = (process.env.HOST as string) || "localhost";
+const FRONTEND_BUILD_LOCATION = "../../frontend/dist";
 
 mongoose.connect(process.env.MONGODB_CONNECTION_STRING as string);
 const db = mongoose.connection;
@@ -53,6 +56,7 @@ startChangeStream();
 
 const app = express();
 
+app.use(express.static(path.join(__dirname, FRONTEND_BUILD_LOCATION)));
 app.use(cookieParser());
 app.use(express.json({ limit: "200mb" }));
 app.use(express.urlencoded({ limit: "200mb", extended: true }));
@@ -73,6 +77,7 @@ app.use("/api/download", downloadRoutes);
 app.use("/api/download", downloadRoutes);
 app.use("/api/notification", notificationRoutes);
 app.use("/api/notification_settings", notificationSettingRoutes);
+app.use("/api/support_ticket", supportTicketRoutes);
 app.use("/config", configRoutes);
 app.use("/autoMail", autoMailRoutess);
 app.use("/backup", backupRoutes);
@@ -80,6 +85,20 @@ app.use("/version", versionRoutes);
 app.use("/healthcheck", (_, res) =>
   res.send("Inventory backend is running...")
 );
+
+// Catch-all route for unmatched URLs
+if (process.env.NODE_ENV !== "development") {
+  app.get("/*", function (req, res) {
+    res.sendFile(
+      path.join(__dirname, `${FRONTEND_BUILD_LOCATION}/index.html`),
+      function (err) {
+        if (err) {
+          res.status(500).send(err);
+        }
+      }
+    );
+  });
+}
 
 const isAtlasDB = () => {
   return process.env.MONGODB_CONNECTION_STRING?.startsWith("mongodb+srv://");