+++ backend/src/system/automail.ts
@@ -259,7 +259,7 @@ router.post(
       });
     } catch (error) {
       console.log(error);
-      return res.status(500).json({ message: "Something went wrong" });
+      return res.status(500).json({ message: "Something went wrong", error });
     }
   }
 );
@@ -277,7 +277,7 @@ router.get(
       return res.status(200).json(existingSettings);
     } catch (error) {
       console.log(error);
-      return res.status(500).json({ message: "Something went wrong" });
+      return res.status(500).json({ message: "Something went wrong", error });
     }
   }
 );
@@ -296,7 +296,7 @@ router.post(
       });
     } catch (error) {
       console.log(error);
-      return res.status(500).json({ message: "Something went wrong" });
+      return res.status(500).json({ message: "Something went wrong", error });
     }
   }
 );