+++ frontend/tailwind.config.cjs
@@ -33,6 +33,7 @@ module.exports = {
           DEFAULT: "var(--brand-green)",
           hover: "var(--brand-green-hover)",
         },
+        brandgreen: "hsla(var(--brand-green))",
         input: "hsl(var(--input))",
         ring: "hsl(var(--ring))",
         tab: "hsl(var(--tab))",